+++ Routers/Routers.js
@@ -2,14 +2,14 @@ const express = require('express')
 const Router = express.Router()
 
 // Import Controller
-const UserController = require('./../Controllers/Controller')
-
+const UserController = require('../Controllers/UserController')
+const AuthController = require('../Controllers/AuthController')
 // Import JWTVerify 
 const jwtVerify = require('./../Middleware/JWTVerify')
 
 // Routing
 Router.post('/register', UserController.register)
-Router.post('/login', UserController.login)
+Router.post('/login', AuthController.login)
 Router.patch('/changePassword', jwtVerify, UserController.changePassword)
 Router.patch('/forgotPassword',  UserController.changePassword)
 Router.get('/profile', jwtVerify, UserController.getUserProfile)
